server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
httpx
PyJWT
passlib
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
httpx
pyjwt
cryptography
pymysql
aiosqlite
